backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
cryptography==42.0.5
Pillow==10.2.0
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
timm==0.9.16
